USE HC158/AHC158s to allow me to change the aux hardware on this board
The issue I'm seeing is that this board is very complex and I need a way
to fix bugs without adding a lot of strange garbage. My plan is to move
the management engine and various parts onto a board that is plugged
into the main board. The other way I could do it is provide this
capability but use multiplexer chips to select the source. That way it
can work without this. Plus I can still fix bugs in the future.

Because of this, the management engine will be responsible for:

- Clock Generation
-- 10MHz is generated externally to save on CCLs
- Ready signal synchronization
-- This is one of the features that will be overloadable through this external connector.
- Fail Circuit
-- Another one of the features that is a combination of internal and
external circuitry
-- Can override this through the exernal connector
- Normal Management Engine Duties
- Handling Reset signal management
-- Keep i960 in reset while waiting for the