Added a 16-bit transceiver to data lines
There is no latch in this design, the GCM4 is responsible for keeping
the lines active when handling read requests from the i960. This isn't
that big of a deal. It just means that we have to keep the data line id
set for as long as possible.

This is basically what I do now for the gcm4 based boards just without
the address bits set. The data lines stay set until the next time they
need to be modified.