Make the management engine responsible for handling reset capabilities
My idea is to use the event system to route the ME_Booted pin to a ccl
within the me to complete the package.

The output of this ccl will be RESET960 and that can be handled manually
until we get the CCLs up. Then we can wait until all the parts have
started up. Hell we can use an AHC138 for this as well!