Use the Management Engine for the reset controller.
It becomes much easier to work around logic bugs with the CCL. I just
setup PF0 as an output but wire that to CCL3.EVTA (I do this for the
	10MHz signal tests where I routed PA7 as a event source)

So that allows me to control the CCL that way. If that fails then I can
just do polling on boot with no problems.

I also added pullups to make sure that on a reset from the chipset we
will keep the i960 in reset.