Put the external ready hardware on a breakout connector
If I run into problems or bugs, we can connect a small handler board
onto the header and override the issue really quickly. If you don't plug
in the connector then no problem we will detect that.

I have it routed through the CCL in a very trivial manner. The nice
thing is that regardless of where we do the ready signal computation we
will always use CCL2.IN0 for ~{ME_READY}